third_party/WebKit/LayoutTests/crypto/ecdh-deriveBits-failures-expected.txt

Tests bad inputs to ECDH's deriveBits()

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


error is: SyntaxError: Algorithm: name: Missing or not a string
error is: SyntaxError: EcdhKeyDeriveParams: public: Missing required property
error is: SyntaxError: Algorithm: Not an object
error is: TypeError: Failed to execute 'deriveBits' on 'SubtleCrypto': parameter 1 ('algorithm') is not an object.
error is: SyntaxError: EcdhKeyDeriveParams: public: Must be a CryptoKey
error is: SyntaxError: EcdhKeyDeriveParams: public: Must be a CryptoKey
error is: SyntaxError: EcdhKeyDeriveParams: public: Must be a CryptoKey
error is: InvalidAccessError: The public parameter for ECDH key derivation is for a different named curve
error is: InvalidAccessError: The public parameter for ECDH key derivation is not a public EC key
error is: InvalidAccessError: The public parameter for ECDH key derivation must be for ECDH
PASS successfullyParsed is true

TEST COMPLETE
